@@ -10,7 +10,7 @@
         }
     ],
     "require": {
-        "php": "^7.4|^8.0",
+        "php": "^8.0",
         "eventsauce/eventsauce": "^1.0",
         "symfony/yaml": "^4.0|^5.0"
     },
